Move REPL engine to a library for reuse
