+++ .idea/runConfigurations/Run_lexer_REPL.xml
@@ -1,10 +1,10 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run lexer REPL" type="CargoCommandRunConfiguration" factoryName="Cargo Command" folderName="lexer" singleton="true">
     <option name="channel" value="DEFAULT" />
-    <option name="command" value="run --package=nafi-lexer-repl" />
+    <option name="command" value="run" />
     <option name="nocapture" value="true" />
     <option name="backtrace" value="SHORT" />
-    <option name="workingDirectory" value="file://$PROJECT_DIR$/lexer" />
+    <option name="workingDirectory" value="file://$PROJECT_DIR$/lexer/repl" />
     <envs />
     <method />
   </configuration>
