+++ lexer/repl/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.0.0"
 publish = false
 
 [dependencies]
-nafi-lexer = { path = "../../lexer/" }
+nafi-lexer = { path = "../" }
 nafi-tokens = { path = "../../tokens/" }
 
 quicli = "0.2"
